UCT/IB_MLX5: code refactoring
code common to rc_mlx5 and ud_mlx5 is moved to ib/mlx5
