+++ src/uct/ib/rc/accel/rc_mlx5_ep.c
@@ -17,15 +17,6 @@
 #define UCT_RC_MLX5_OPCODE_MASK       0xff
 
 
-static inline void uct_rc_mlx5_bf_copy_bb(void *dst, void *src)
-{
-#ifdef __SSE4_2__
-    UCS_WORD_COPY(dst, src, __m128i, MLX5_SEND_WQE_BB);
-#else
-    UCS_WORD_COPY(dst, src, uint64_t, MLX5_SEND_WQE_BB);
-#endif
-}
-
 static UCS_F_ALWAYS_INLINE void
 uct_rc_mlx5_ep_set_rdma_seg(struct mlx5_wqe_raddr_seg *raddr, uint64_t rdma_raddr,
                             uct_rkey_t rdma_rkey)
@@ -43,95 +34,17 @@ uct_rc_mlx5_ep_set_rdma_seg(struct mlx5_wqe_raddr_seg *raddr, uint64_t rdma_radd
 #endif
 }
 
-static UCS_F_ALWAYS_INLINE
-void uct_rc_mlx5_ep_set_dptr_seg(struct mlx5_wqe_data_seg *dptr, const void *address,
-                                 unsigned length, uint32_t lkey)
-{
-    dptr->byte_count = htonl(length);
-    dptr->lkey       = htonl(lkey);
-    dptr->addr       = htonll((uintptr_t)address);
-}
-
-void uct_rc_mlx5_set_ctrl_seg(struct mlx5_wqe_ctrl_seg* ctrl, uint16_t pi,
-                              uint8_t opcode, uint8_t opmod, uint32_t qp_num,
-                              uint8_t fm_ce_se, uint8_t ds)
-{
-#ifdef __SSE4_2__
-    *(__m128i *) ctrl = _mm_shuffle_epi8(
-                    _mm_set_epi32(qp_num, ds, pi,
-                                  (opcode << 16) | (opmod << 8) | fm_ce_se), /* OR of constants */
-                    _mm_set_epi8(0, 0, 0, 0, /* immediate */
-                                 0,          /* signal/fence_mode */
-                                 0, 0,       /* reserved */
-                                 0,          /* signature */
-                                 8,          /* data size */
-                                 12, 13, 14, /* QP num */
-                                 2,          /* opcode */
-                                 4, 5,       /* sw_pi in BE */
-                                 1           /* opmod */
-                                 ));
-#else
-    ctrl->opmod_idx_opcode = (opcode << 24) | (htons(pi) << 8) | opmod;
-    ctrl->qpn_ds           = htonl((qp_num << 8) | ds);
-    ctrl->fm_ce_se         = fm_ce_se;
-#endif
-}
-
 static UCS_F_ALWAYS_INLINE void
 uct_rc_mlx5_post_send(uct_rc_mlx5_ep_t *ep, struct mlx5_wqe_ctrl_seg *ctrl,
                       uint8_t opcode, uint8_t opmod, unsigned sig_flag, unsigned wqe_size)
 {
-    unsigned n, num_seg, num_bb;
-    void *src, *dst;
-    uint16_t sw_pi;
+    uct_ib_mlx5_set_ctrl_seg(ctrl, ep->tx.wq.sw_pi, 
+                             opcode, opmod, ep->qp_num, sig_flag, wqe_size);
 
-    num_seg = ucs_div_round_up(wqe_size, UCT_IB_MLX5_WQE_SEG_SIZE);
-    num_bb  = ucs_div_round_up(wqe_size, MLX5_SEND_WQE_BB);
-    sw_pi   = ep->tx.sw_pi;
-
-    uct_rc_mlx5_set_ctrl_seg(ctrl, sw_pi, opcode, opmod, ep->qp_num, sig_flag,
-                             num_seg);
-    uct_ib_mlx5_log_tx(IBV_QPT_RC, ctrl, ep->tx.qstart, ep->tx.qend,
+    uct_ib_mlx5_log_tx(IBV_QPT_RC, ctrl, ep->tx.wq.qstart, ep->tx.wq.qend,
                        (opcode == MLX5_OPCODE_SEND) ? uct_rc_ep_am_packet_dump : NULL);
 
-    /* TODO Put memory store fence here too, to prevent WC being flushed after DBrec */
-    ucs_memory_cpu_store_fence();
-
-    /* Write doorbell record */
-    ep->tx.prev_sw_pi = sw_pi;
-    *ep->tx.dbrec = htonl(sw_pi += num_bb);
-
-    /* Make sure that doorbell record is written before ringing the doorbell */
-    ucs_memory_bus_store_fence();
-
-    /* Set up copy pointers */
-    dst = ep->tx.bf_reg;
-    src = ctrl;
-
-    /* BF copy */
-    /* TODO support DB without BF */
-    ucs_assert(wqe_size <= ep->tx.bf_size);
-    ucs_assert(num_bb <= UCT_RC_MLX5_MAX_BB);
-    for (n = 0; n < num_bb; ++n) {
-        uct_rc_mlx5_bf_copy_bb(dst, src);
-        dst += MLX5_SEND_WQE_BB;
-        src += MLX5_SEND_WQE_BB;
-        if (ucs_unlikely(src == ep->tx.qend)) {
-            src = ep->tx.qstart;
-        }
-    }
-
-    /* We don't want the compiler to reorder instructions and hurt latency */
-    ucs_compiler_fence();
-
-    /* Advance queue pointer */
-    ucs_assert(ctrl == ep->tx.seg);
-    ep->tx.seg   = src;
-    ep->tx.sw_pi = sw_pi;
-
-    /* Flip BF register */
-    ep->tx.bf_reg = (void*) ((uintptr_t) ep->tx.bf_reg ^ ep->tx.bf_size);
-
+    uct_ib_mlx5_post_send(&ep->tx.wq, ctrl, wqe_size);
     uct_rc_ep_tx_posted(&ep->super, sig_flag & MLX5_WQE_CTRL_CQ_UPDATE);
 }
 
@@ -142,29 +55,6 @@ uct_rc_mlx5_iface_tx_moderation(uct_rc_mlx5_ep_t* ep, uct_rc_mlx5_iface_t* iface
            MLX5_WQE_CTRL_CQ_UPDATE : 0;
 }
 
-/**
- * Copy data to inline segment, taking into account QP wrap-around.
- *
- * @param dest    Inline data in the WQE to copy to.
- * @patam src     Data to copy.
- * @param length  Data length.
- *
- * @return If there was a wrap-around, return -qp_size. Otherwise, return 0.
- */
-static UCS_F_ALWAYS_INLINE void
-uct_rc_mlx5_inline_copy(void *dest, const void *src, unsigned length, uct_rc_mlx5_ep_t *ep)
-{
-    void *qend = ep->tx.qend;
-    ptrdiff_t n;
-
-    if (dest + length <= qend) {
-        memcpy(dest, src, length);
-    } else {
-        n = qend - dest;
-        memcpy(dest, src, n);
-        memcpy(ep->tx.qstart, src + n, length - n);
-    }
-}
 
 /*
  * Generic inline posting function.
@@ -194,7 +84,7 @@ uct_rc_mlx5_ep_inline_post(uct_rc_mlx5_ep_t *ep, unsigned opcode,
     unsigned wqe_size;
     unsigned sig_flag;
 
-    ctrl = ep->tx.seg;
+    ctrl = ep->tx.wq.seg;
     UCT_RC_MLX5_CHECK_RES(iface, ep);
 
     switch (opcode) {
@@ -208,7 +98,7 @@ uct_rc_mlx5_ep_inline_post(uct_rc_mlx5_ep_t *ep, unsigned opcode,
         am               = (void*)(inl + 1);
         am->rc_hdr.am_id = am_id;
         am->am_hdr       = am_hdr;
-        uct_rc_mlx5_inline_copy(am + 1, buffer, length, ep);
+        uct_ib_mlx5_inline_copy(am + 1, buffer, length, &ep->tx.wq);
         sig_flag         = uct_rc_iface_tx_moderation(&iface->super, &ep->super,
                                                       MLX5_WQE_CTRL_CQ_UPDATE);
         break;
@@ -226,7 +116,7 @@ uct_rc_mlx5_ep_inline_post(uct_rc_mlx5_ep_t *ep, unsigned opcode,
         uct_rc_mlx5_ep_set_rdma_seg(raddr, rdma_raddr, rdma_rkey);
         inl              = (void*)(raddr + 1);
         inl->byte_count  = htonl(length | MLX5_INLINE_SEG);
-        uct_rc_mlx5_inline_copy(inl + 1, buffer, length, ep);
+        uct_ib_mlx5_inline_copy(inl + 1, buffer, length, &ep->tx.wq);
         sig_flag         = MLX5_WQE_CTRL_CQ_UPDATE;
         break;
 
@@ -289,7 +179,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
     }
 
     opmod = 0;
-    ctrl = ep->tx.seg;
+    ctrl = ep->tx.wq.seg;
     switch (opcode_flags) {
     case MLX5_OPCODE_SEND:
         UCT_CHECK_LENGTH(length + sizeof(*rch) + am_hdr_len,
@@ -306,7 +196,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
         rch              = (void*)(inl + 1);
         rch->am_id       = am_id;
 
-        uct_rc_mlx5_inline_copy(rch + 1, am_hdr, am_hdr_len, ep);
+        uct_ib_mlx5_inline_copy(rch + 1, am_hdr, am_hdr_len, &ep->tx.wq);
 
         /* Data segment with payload */
         if (length == 0) {
@@ -314,13 +204,13 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
         } else {
             wqe_size     = sizeof(*ctrl) + inl_seg_size + sizeof(*dptr);
             dptr         = (void*)(ctrl + 1) + inl_seg_size;
-            if (ucs_unlikely((void*)dptr >= ep->tx.qend)) {
-                dptr = (void*)dptr - (ep->tx.qend - ep->tx.qstart);
+            if (ucs_unlikely((void*)dptr >= ep->tx.wq.qend)) {
+                dptr = (void*)dptr - (ep->tx.wq.qend - ep->tx.wq.qstart);
             }
 
-            ucs_assert((void*)dptr       >= ep->tx.qstart);
-            ucs_assert((void*)(dptr + 1) <= ep->tx.qend);
-            uct_rc_mlx5_ep_set_dptr_seg(dptr, buffer, length, *lkey_p);
+            ucs_assert((void*)dptr       >= ep->tx.wq.qstart);
+            ucs_assert((void*)(dptr + 1) <= ep->tx.wq.qend);
+            uct_ib_mlx5_set_data_seg(dptr, buffer, length, *lkey_p);
         }
         break;
 
@@ -331,7 +221,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
         ucs_assert(length < (2ul << 30));
 
         wqe_size         = sizeof(*ctrl) + sizeof(*dptr);
-        uct_rc_mlx5_ep_set_dptr_seg((void*)(ctrl + 1), buffer, length, *lkey_p);
+        uct_ib_mlx5_set_data_seg((void*)(ctrl + 1), buffer, length, *lkey_p);
         break;
 
     case MLX5_OPCODE_RDMA_READ:
@@ -347,7 +237,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
             wqe_size     = sizeof(*ctrl) + sizeof(*raddr);
         } else {
             wqe_size     = sizeof(*ctrl) + sizeof(*raddr) + sizeof(*dptr);
-            uct_rc_mlx5_ep_set_dptr_seg((void*)(raddr + 1), buffer, length, *lkey_p);
+            uct_ib_mlx5_set_data_seg((void*)(raddr + 1), buffer, length, *lkey_p);
         }
         break;
 
@@ -363,7 +253,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
         }
         atomic->swap_add  = swap_add;
 
-        uct_rc_mlx5_ep_set_dptr_seg((void*)(atomic + 1), buffer, length, *lkey_p);
+        uct_ib_mlx5_set_data_seg((void*)(atomic + 1), buffer, length, *lkey_p);
         wqe_size          = sizeof(*ctrl) + sizeof(*raddr) + sizeof(*atomic) +
                             sizeof(*dptr);
         break;
@@ -398,18 +288,18 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
             /* Handle QP wrap-around. It cannot happen in the middle of
              * masked-cswap segment, because it's still in the first BB.
              */
-            ucs_assert((void*)dptr <= ep->tx.qend);
-            if (dptr == ep->tx.qend) {
-                dptr = ep->tx.qstart;
+            ucs_assert((void*)dptr <= ep->tx.wq.qend);
+            if (dptr == ep->tx.wq.qend) {
+                dptr = ep->tx.wq.qstart;
             } else {
-                ucs_assert((void*)masked_cswap64 < ep->tx.qend);
+                ucs_assert((void*)masked_cswap64 < ep->tx.wq.qend);
             }
             break;
         default:
             ucs_assert(0);
         }
 
-        uct_rc_mlx5_ep_set_dptr_seg(dptr, buffer, length, *lkey_p);
+        uct_ib_mlx5_set_data_seg(dptr, buffer, length, *lkey_p);
         break;
 
      case MLX5_OPCODE_ATOMIC_MASKED_FA:
@@ -422,7 +312,7 @@ uct_rc_mlx5_ep_dptr_post(uct_rc_mlx5_ep_t *ep, unsigned opcode_flags,
         masked_fadd32->add            = swap_add;
         masked_fadd32->filed_boundary = 0;
 
-        uct_rc_mlx5_ep_set_dptr_seg((void*)(masked_fadd32 + 1), buffer, length,
+        uct_ib_mlx5_set_data_seg((void*)(masked_fadd32 + 1), buffer, length,
                                     *lkey_p);
         wqe_size                      = sizeof(*ctrl) + sizeof(*raddr) +
                                         sizeof(*masked_fadd32) + sizeof(*dptr);
@@ -451,7 +341,7 @@ uct_rc_mlx5_ep_bcopy_post(uct_rc_mlx5_ep_t *ep, unsigned opcode, unsigned length
 {
     ucs_status_t status;
 
-    desc->super.sn = ep->tx.sw_pi;
+    desc->super.sn = ep->tx.wq.sw_pi;
     status = uct_rc_mlx5_ep_dptr_post(ep, opcode, desc + 1, length, &desc->lkey,
                                       am_id, am_hdr, am_hdr_len, rdma_raddr, rdma_rkey,
                                       0, 0, 0, force_sig);
@@ -482,7 +372,7 @@ uct_rc_mlx5_ep_zcopy_post(uct_rc_mlx5_ep_t *ep, unsigned opcode, const void *buf
 
     UCT_RC_MLX5_CHECK_RES(iface, ep);
 
-    sn = ep->tx.sw_pi;
+    sn = ep->tx.wq.sw_pi;
     status = uct_rc_mlx5_ep_dptr_post(ep, opcode, buffer, length, &mr->lkey,
                                       am_id, am_hdr, am_hdr_len, rdma_raddr, rdma_rkey,
                                       0, 0, 0,
@@ -504,7 +394,7 @@ uct_rc_mlx5_ep_atomic_post(uct_rc_mlx5_ep_t *ep, unsigned opcode,
 {
     ucs_status_t status;
 
-    desc->super.sn = ep->tx.sw_pi;
+    desc->super.sn = ep->tx.wq.sw_pi;
     status = uct_rc_mlx5_ep_dptr_post(ep, opcode, desc + 1, length, &desc->lkey,
                                       0, NULL, 0, remote_addr, rkey, compare_mask,
                                       compare, swap_add, signal);
@@ -792,7 +682,7 @@ ucs_status_t uct_rc_mlx5_ep_flush(uct_ep_h tl_ep)
      * to the value calculated from prev_sw_pi - which is the index where the last
      * posted WQE started. See also uct_rc_mlx5_iface_poll_tx().
      */
-    exp_max_pi = uct_rc_mlx5_calc_max_pi(iface, ep->tx.prev_sw_pi);
+    exp_max_pi = uct_rc_mlx5_calc_max_pi(iface, ep->tx.wq.prev_sw_pi);
     if (ep->tx.max_pi == exp_max_pi) {
         UCT_TL_EP_STAT_FLUSH(&ep->super.super);
         ucs_trace_data("ep %p is flushed", ep);
@@ -813,39 +703,18 @@ ucs_status_t uct_rc_mlx5_ep_flush(uct_ep_h tl_ep)
 static UCS_CLASS_INIT_FUNC(uct_rc_mlx5_ep_t, uct_iface_h tl_iface)
 {
     uct_rc_mlx5_iface_t *iface = ucs_derived_of(tl_iface, uct_rc_mlx5_iface_t);
-    uct_ib_mlx5_qp_info_t qp_info;
     ucs_status_t status;
 
     UCS_CLASS_CALL_SUPER_INIT(uct_rc_ep_t, &iface->super);
 
-
-    /* TODO: must be common code with UD */
-    status = uct_ib_mlx5_get_qp_info(self->super.qp, &qp_info);
+    status = uct_ib_mlx5_get_txwq(self->super.qp, &self->tx.wq);
     if (status != UCS_OK) {
         ucs_error("Failed to get mlx5 QP information");
         return status;
     }
 
-    if ((qp_info.bf.size == 0) || !ucs_is_pow2(qp_info.bf.size) ||
-        (qp_info.sq.stride != MLX5_SEND_WQE_BB) ||
-        !ucs_is_pow2(qp_info.sq.wqe_cnt))
-    {
-        ucs_error("mlx5 device parameters not suitable for transport");
-        return UCS_ERR_IO_ERROR;
-    }
-
-    self->qp_num        = self->super.qp->qp_num;
-    self->tx.qstart     = qp_info.sq.buf;
-    self->tx.qend       = qp_info.sq.buf + (MLX5_SEND_WQE_BB *  qp_info.sq.wqe_cnt);
-    self->tx.seg        = self->tx.qstart;
-    self->tx.sw_pi      = 0;
-    self->tx.prev_sw_pi = -1;
-    self->tx.max_pi     = uct_rc_mlx5_calc_max_pi(iface, self->tx.prev_sw_pi);
-    self->tx.bf_reg     = qp_info.bf.reg;
-    self->tx.bf_size    = qp_info.bf.size;
-    self->tx.dbrec      = &qp_info.dbrec[MLX5_SND_DBR];
-
-    memset(self->tx.qstart, 0, self->tx.qend - self->tx.qstart);
+    self->qp_num    = self->super.qp->qp_num;
+    self->tx.max_pi = uct_rc_mlx5_calc_max_pi(iface, self->tx.wq.prev_sw_pi);
     return UCS_OK;
 }
